requirements.txt
# Portkey Integration
portkey-ai==1.6.0

# Data Processing
pandas==2.1.4

# Dashboard
streamlit==1.30.0
plotly==5.18.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
